+++ src/packages/relay/src/compiler/__tests__/getSchemaSource.test.js
@@ -2,21 +2,25 @@
 
 import fs from 'fs';
 import path from 'path';
-import { GraphQLSchema } from 'graphql';
+import { Source } from 'graphql';
 
-import getSchema from '../getSchema';
+import getSchemaSource from '../getSchemaSource';
 
 const basePath = path.join(__dirname, '__fixtures__', 'schemas');
 test.each(fs.readdirSync(basePath))('loads the fixture as expected', fixturePath => {
-  expect.hasAssertions();
-  let schema;
+  expect.assertions(1);
+
+  let schemaSource;
   const schemaPath = path.join(basePath, fixturePath);
   try {
-    schema = getSchema(schemaPath);
-    expect(schema instanceof GraphQLSchema).toBe(true);
+    schemaSource = getSchemaSource(schemaPath);
   } catch (error) {
     expect(
       `\n${error.message.replace(schemaPath, path.relative(basePath, schemaPath))}\n`,
     ).toMatchSnapshot();
+  }
+
+  if (schemaSource) {
+    expect(schemaSource instanceof Source).toBe(true);
   }
 });