+++ src/packages/relay/src/createLocalEnvironment.js
@@ -1,11 +1,10 @@
 // @flow
 
-import { RecordSource, Store, Environment as RelayEnvironment } from 'relay-runtime';
+import { Environment as RelayEnvironment } from 'relay-runtime';
 
+import createRelayStore from './createRelayStore';
+import RelayLogger from './RelayLogger';
 import type { Environment } from './runtimeTypes.flow';
-
-const source = new RecordSource();
-const store = new Store(source);
 
 /**
  * This is just an alternative environment factory for LocalQueryRenderer. There is currently not
@@ -14,6 +13,7 @@ const store = new Store(source);
 export default function createLocalEnvironment(): Environment {
   return new RelayEnvironment({
     // notice this environment doesn't have network layer since it's not necessary
-    store,
+    log: RelayLogger,
+    store: createRelayStore(),
   });
 }